Cut assembly peak RSS in half again, with byte-identical output
Peak RSS on yeast_sub 4,200 -> 3,690 MB, and 7,492 MB at the start of this
session's memory work. Every change here is verified byte-identical on a
deterministic run (same size, same md5), not merely lossless: holding the
archive exactly is the constraint, so nothing that costs bytes qualifies --
which rules out the hash APSP (+0.31%), the 2-bit pg codec (+5.9%) and the
FCM table cap, all of which buy memory or speed with archive size.

Instrumenting rather than reasoning found the peak. ARCS_VODBG_MEM=1 reports
current RSS at each allocation point, and SA-MEM does the same inside the
suffix-array build, where the caller's marks could not see:

    build entry     1145 MB
    sa allocated    2126 MB   +981
    lcp allocated   3107 MB   +981
    plcp allocated  4093 MB   +981   <- the actual peak, three int32 arrays
    plcp freed      3111 MB

Four changes:

both_views held a third copy of every sequence -- forward already in
orig_reads, reverse-complement in rcseq -- for builders that only read
.data()/.size(). Now string_views. Freeing the copies was never an option:
releasing 312 MB returned 52 MB, because glibc keeps the arena.

The APSP candidate table is read only by growth, but stayed live through the
fallback k-mer index build, which is where peak RSS actually lands (2405,
2407, 2964 MB across the three trials). Released after the final trial's
growth, with malloc_trim so the arena actually returns: 1755 -> 1090 MB.

rcseq was 851k separate string allocations -- a 32-byte header plus its own
heap block and rounding, for a ~150-byte payload. Now one arena with an offset
table, with record_mapped/record_append/pack_kmer widened to string_view so
nothing downstream needs a materialised string.

LCP is uint16, not uint32. Reads are separated by a byte no base can equal, so
these values sit in read-length territory (~150) where a uint32 reserves four
bytes; saturating at 65535 is lossless given overlaps are capped by read
length and chain-pg refuses reads beyond 65535 bp.

NEGATIVE RESULT, recorded in sa_apsp.cpp: PLCP can be removed entirely by
permuting it into LCP in place, since LCP[i] = PLCP[SA[i]] is a gather along a
permutation. It works and is byte-identical, and it drops the peak further to
3,111 MB -- but cycle-following chases SA through ~257M random accesses,
serially, twice per run, and wall time went 20 -> 74 s. Rejected: the goal is
memory AND speed. A parallel gather into the narrowed array recovers most of
the memory for none of the time.

Also: the exact-size CSR k-mer table now has its own switch (ARCS_CSR_INDEX)
and is measured size-neutral on the default path (-97 bytes), so it can be
adopted for the overflow-drop fix on its own merits.

8/8 ctest; yeast_sub, byte-identical archives at every step.

+++ src/repeat_elim.cpp
@@ -118,7 +118,7 @@ bool repeat_elim_encode_cap(const std::string& pg, uint32_t K,
             uint32_t running = UINT32_MAX;
             uint32_t steps = 0;
             for (int64_t rr = (int64_t)r - 1; rr >= 0 && steps < search_cap; --rr, ++steps) {
-                running = std::min(running, SA.lcp[(size_t)(rr + 1)]);
+                running = std::min(running, (uint32_t)SA.lcp[(size_t)(rr + 1)]);
                 if (running < K) break;
                 uint32_t p = SA.sa[(size_t)rr];
                 if (p < (uint32_t)n) consider_forward(p, running);
@@ -127,7 +127,7 @@ bool repeat_elim_encode_cap(const std::string& pg, uint32_t K,
             running = UINT32_MAX;
             steps = 0;
             for (size_t rr = (size_t)r + 1; rr < Tn && steps < search_cap; ++rr, ++steps) {
-                running = std::min(running, SA.lcp[rr]);
+                running = std::min(running, (uint32_t)SA.lcp[rr]);
                 if (running < K) break;
                 uint32_t p = SA.sa[rr];
                 if (p < (uint32_t)n) consider_forward(p, running);
